+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+ - Removed table alias in `ForeignKeyAfterTask` for **PostgreSQL** foreign keys `UPDATE` query
 
 ## [1.0.2] - 2021-05-22
 ### Fixed